+++ R/utils.R
@@ -1,3 +1,5 @@
+#' Set global options for [roboplot()]
+#'
 #' Override the default options for colors, fonts etc. for any plot created with [roboplot()]
 #'
 #' @param accessible Logical. Forces trace colors accessible against plot background.
@@ -189,7 +191,9 @@
 #' # parameter 'shinyapp' as a list with a character string named
 #' # 'container' describing the css selector of the container that controls plot
 #' # visibility allows roboplotr::roboplot() to detect if the container was
-#' # displayed on plot rendering, redoing the incorrect layouts.
+#' # displayed on plot rendering, redoing the incorrect layouts. The example
+#' # below has one card starting as collapsed as an example of use case where
+#' # this does not work at all.
 #'
 #' if(interactive()) {
 #'
@@ -208,8 +212,7 @@
 #'       )
 #'     ),
 #'     bs4Dash::dashboardBody(
-#'       # Boxes need to be put in a row (or column)
-#'       uiOutput("plots")
+#'       shiny::uiOutput("plots")
 #'     )
 #'   )
 #'
@@ -218,25 +221,26 @@
 #'     output$plots <- renderUI({
 #'       bs4Dash::tabItems(
 #'         bs4Dash::tabItem(tabName = "tab_1", shiny::fluidRow(
-#'           bs4Dash::box(plotly::plotlyOutput("plot1", height = 550))
+#'           bs4Dash::box(plotly::plotlyOutput("plot1", height = 550), collapsed = T,
+#'                        container = ".card-body")
 #'         )),
 #'         bs4Dash::tabItem(tabName = "tab_2", shiny::fluidRow(
-#'           bs4Dash::box(plotly::plotlyOutput("plot2", height = 500))
+#'           bs4Dash::box(plotly::plotlyOutput("plot2", height = 500),collapsed = F)
 #'         ))
 #'       )
 #'     })
 #'
 #'     output$plot1 <- plotly::renderPlotly({
 #'       roboplot(
-#'         dplyr::filter(energiantuonti, Suunta == "Tuonti", Alue == "Kanada"),
+#'         dplyr::filter(energiantuonti, Suunta == "Tuonti"),
 #'         Alue, "Energian tuonti Kanadasta",
 #'         "Milj. €",
 #'         "Tilastokeskus")
 #'     })
 #'     output$plot2 <- plotly::renderPlotly({
 #'       roboplot(
-#'         dplyr::filter(energiantuonti, Suunta == "Vienti", Alue == "Kanada"),
-#'         Alue, "Energian vienti Kanadaan",
+#'         dplyr::filter(energiantuonti, Suunta == "Vienti"),
+#'         Alue, "Energian vienti",
 #'         "Milj. €",
 #'         "Tilastokeskus",
 #'         height = 500)
@@ -248,6 +252,7 @@
 #'   shiny::shinyApp(ui, server)
 #'
 #' }
+#'
 #'
 #' @importFrom htmltools singleton tagList tags
 #' @importFrom purrr iwalk
@@ -471,6 +476,8 @@ roboplotr_string2filename <- function(string) {
     str_replace_all(c("\uE4" = "a", "\uE5" = "o", "\uF6" = "o", " |\\.|-" = "_", "," = ""))
 }
 
+#' Locales for [roboplot()]
+#'
 #' Used to set locale parameters for [roboplot()] in [set_roboplot_options()]
 #'
 #' @param locale Character. Currently supports on only "en-GB", "en-US",